+++ fabbwled-frontend/src/routes/create/chooseCreate/choose-create.scss
@@ -35,7 +35,7 @@
 .choose_create_book {
   height: 80vh;
   width: 80vw;
-  background-image: url("../../assets/book.png");
+  background-image: url("../../../assets/book.png");
   background-size: cover;
   text-align: center;
   display: flex;
